+++ data/src/main/java/com/proptiger/data/model/Suburb.java
@@ -8,7 +8,6 @@ import javax.persistence.Entity;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
-import javax.persistence.MapsId;
 import javax.persistence.Table;
 
 import com.proptiger.data.meta.FieldMetaInfo;
@@ -25,11 +24,11 @@ public class Suburb {
     @Id
     @FieldMetaInfo( displayName = "Suburb Id",  description = "Suburb Id")
     @Column(name = "SUBURB_ID")
-    private long id;
+    private int id;
 
     @FieldMetaInfo( displayName = "City Id",  description = "City Id")
     @Column(name = "CITY_ID")
-    private long cityId;
+    private int cityId;
 
     @FieldMetaInfo( displayName = "Label",  description = "Suburb label")
     @Column(name = "LABEL")
@@ -39,11 +38,11 @@ public class Suburb {
     @JoinColumn(name="CITY_ID", insertable = false, updatable = false)
     private City city;
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
 
@@ -63,11 +62,11 @@ public class Suburb {
         this.city = city;
     }
 
-    public long getCityId() {
+    public int getCityId() {
         return cityId;
     }
 
-    public void setCityId(long cityId) {
+    public void setCityId(int cityId) {
         this.cityId = cityId;
     }
 }
